+++ src/io/+mag/+io/+in/Format.m
@@ -1,6 +1,11 @@
 classdef (Abstract) Format < matlab.mixin.Heterogeneous & mag.mixin.SetGet
 % IFORMAT Interface for data format providers for import.
 
+    properties (Abstract, Constant)
+        % EXTENSION Extension supported for file format.
+        Extension (1, 1) string
+    end
+
     methods (Abstract)
 
         % LOAD Load raw data from file.