+++ src/io/+mag/+io/+in/ScienceCSV.m
@@ -97,7 +97,7 @@ classdef ScienceCSV < mag.io.in.CSV
             rawData.quality = repmat(mag.meta.Quality.Regular, height(rawData), 1);
 
             % Convert timestamps.
-            for ps = [mag.process.Timestamp(), mag.process.DateTime()]
+            for ps = [mag.process.Missing(Variables = ["x", "y", "z"]), mag.process.Timestamp(), mag.process.DateTime()]
                 rawData = ps.apply(rawData, metaData);
             end