+++ resources/ReleaseNotes.md
@@ -1,10 +1,13 @@
 # App
 
+- Adapt to refactoring of import/export
 - Add support for CDF import/export
 - Remove support for DAT export
 
 # Software
 
 - Refactor import/export of science and HK data
-- Add support for CDF import/export
+- Add support for CDF science import/export
+- Add support in `mag.meta.log.Word` for EM documents
 - Remove support for DAT export
+- Move description of sensor setup to `mag.meta.Setup`